调整模块项目结构
