+++ growingio-autotracker-core/src/main/java/com/growingio/android/sdk/autotrack/inject/UcWebViewInjector.java
@@ -19,7 +19,7 @@ package com.growingio.android.sdk.autotrack.inject;
 import com.growingio.android.sdk.TrackerContext;
 import com.growingio.android.sdk.track.log.Logger;
 import com.growingio.android.sdk.track.modelloader.ModelLoader;
-import com.growingio.android.sdk.track.modelloader.data.HybridBridge;
+import com.growingio.android.sdk.track.middleware.hybrid.HybridBridge;
 import com.uc.webview.export.WebView;
 
 import java.util.Map;
