+++ growingio-autotracker-core/src/main/java/com/growingio/android/sdk/autotrack/inject/X5WebViewInjector.java
@@ -20,7 +20,7 @@ package com.growingio.android.sdk.autotrack.inject;
 import com.growingio.android.sdk.TrackerContext;
 import com.growingio.android.sdk.track.log.Logger;
 import com.growingio.android.sdk.track.modelloader.ModelLoader;
-import com.growingio.android.sdk.track.modelloader.data.HybridBridge;
+import com.growingio.android.sdk.track.middleware.hybrid.HybridBridge;
 import com.tencent.smtt.sdk.WebView;
 
 import java.util.Map;
