+++ growingio-hybrid/src/main/java/com/growingio/android/hybrid/HybridDomLoader.java
@@ -23,8 +23,8 @@ import com.growingio.android.sdk.track.async.Callback;
 import com.growingio.android.sdk.track.modelloader.LoadDataFetcher;
 import com.growingio.android.sdk.track.modelloader.ModelLoader;
 import com.growingio.android.sdk.track.modelloader.ModelLoaderFactory;
-import com.growingio.android.sdk.track.modelloader.data.HybridDom;
-import com.growingio.android.sdk.track.modelloader.data.HybridJson;
+import com.growingio.android.sdk.track.middleware.hybrid.HybridDom;
+import com.growingio.android.sdk.track.middleware.hybrid.HybridJson;
 import com.growingio.android.sdk.track.utils.ClassExistHelper;
 
 import org.json.JSONObject;
