+++ growingio-network/urlconnection/src/main/java/com/growingio/android/urlconnection/UrlConnectionFetcher.java
@@ -18,8 +18,8 @@ package com.growingio.android.urlconnection;
 
 import android.text.TextUtils;
 
-import com.growingio.android.sdk.track.http.EventResponse;
-import com.growingio.android.sdk.track.http.EventUrl;
+import com.growingio.android.sdk.track.middleware.http.EventResponse;
+import com.growingio.android.sdk.track.middleware.http.EventUrl;
 import com.growingio.android.sdk.track.log.Logger;
 import com.growingio.android.sdk.track.middleware.http.HttpDataFetcher;
