+++ growingio-tracker-core/src/main/java/com/growingio/android/sdk/track/middleware/hybrid/HybridBridge.java
@@ -1,0 +1,32 @@
+/*
+ *   Copyright (C) 2020 Beijing Yishu Technology Co., Ltd.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+package com.growingio.android.sdk.track.middleware.hybrid;
+
+
+import android.view.View;
+
+public class HybridBridge {
+    private final View view;
+
+    public HybridBridge(View view) {
+        this.view = view;
+    }
+
+    public View getView() {
+        return view;
+    }
+}
