+++ growingio-webservice/circler/src/main/java/com/growingio/android/circler/screenshot/CircleScreenshot.java
@@ -30,8 +30,8 @@ import com.growingio.android.sdk.autotrack.view.ViewUtil;
 import com.growingio.android.sdk.track.async.Callback;
 import com.growingio.android.sdk.track.modelloader.LoadDataFetcher;
 import com.growingio.android.sdk.track.modelloader.ModelLoader;
-import com.growingio.android.sdk.track.modelloader.data.HybridDom;
-import com.growingio.android.sdk.track.modelloader.data.HybridJson;
+import com.growingio.android.sdk.track.middleware.hybrid.HybridDom;
+import com.growingio.android.sdk.track.middleware.hybrid.HybridJson;
 import com.growingio.android.sdk.track.utils.ClassExistHelper;
 import com.growingio.android.sdk.track.utils.DeviceUtil;
 import com.growingio.android.sdk.track.view.DecorView;
